requirements.txt
# Database
psycopg2-binary>=2.9.9

# Requests & Networking
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.7.22

# Utils
python-dotenv>=1.0.0
configparser>=6.0.0
loguru>=0.7.2
python-dateutil>=2.8.2
tqdm>=4.66.1
rich>=13.7.0

# XML & HTML Parsing
lxml>=4.9.3
beautifulsoup4>=4.12.0

# Document Parsing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
pdf2image>=1.16.3
pytesseract>=0.3.10
# Note: pdf2image requires poppler installed
# Note: pytesseract requires tesseract-ocr installed
